// types/leads.types.ts
export type LEADS_TYPES_RESPONSE = {
  _id: string;
  count: number;
};

export type LEADS_TYPES_MODEL = {
  _id: string;
  count: number;
};

export type LEADS_SOURCE_RESPONSE = {
  _id: string;
  firstName: string;
  lastName: string;
  email: string;
  query: string;
  sourceName: string;
  phoneNumber: string;
  createdAt: string;
  updatedAt: string;
};

export type LEADS_SOURCE_MODEL = {
  _id: string;
  firstName: string;
  lastName: string;
  email: string;
  query: string;
  sourceName: string;
  phoneNumber: string;
  createdAt: string;
  updatedAt: string;
};